Don't use string concat with multivar string interpolation (#49)
Before this change, the module would fail with "Not all arguments
converted string formatting" instead of the error messages it was
trying to produce. This caused a lot of confusion for the user since
there was no way for him/her to diagnose it unless he/she dives into
the dimensiondata_compute module code.

Removing the '+' sign (which is not needed for string concat inside
parentheses anyway) solves the issue and produces the intended
error message.